refactor: bet placement functions to method
