+++ x/market/types/market.go
@@ -45,3 +45,14 @@ func (m *Market) isActiveOrInactive() bool {
 	return m.Status == MarketStatus_MARKET_STATUS_ACTIVE ||
 		m.Status == MarketStatus_MARKET_STATUS_INACTIVE
 }
+
+// HasOdds determine if the inpur odds uid is present in
+// the market odds or not.
+func (m *Market) HasOdds(oddsUID string) bool {
+	for _, o := range m.Odds {
+		if oddsUID == o.UID {
+			return true
+		}
+	}
+	return false
+}
